move auth logic into the Authentication.tsx file
